requirements.txt
groq>=0.4.0
PyQt6>=6.6.0
pynput>=1.7.6
pywinauto>=0.6.8
pyperclip>=1.8.2
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
